requirements.txt
aiosmtplib==2.0.1
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
blinker==1.5
click==8.1.3
databases==0.5.5
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
fastapi-mail
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
ormar==0.12.1
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==1.4.37
# sqlalchemy-migrate=0.13.0
typing_extensions==4.5.0
uvicorn==0.20.0
websockets==10.4
